requirements.txt
robotframework~=3.2.2
PyYAML~=5.3.1
robotframework-requests~=0.7.2
prometheus_client~=0.8.0
